[chore/#112] seatRecord 액티비티 네이밍 수정
