+++ presentation/src/main/java/com/dpm/presentation/seatrecord/SeatRecordActivity.kt
@@ -5,34 +5,35 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import android.view.View.GONE
-import android.view.View.INVISIBLE
 import android.view.View.VISIBLE
+import android.widget.AdapterView
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
+import androidx.core.widget.NestedScrollView
 import androidx.fragment.app.commit
 import androidx.lifecycle.asLiveData
-import androidx.recyclerview.widget.RecyclerView
-import androidx.recyclerview.widget.RecyclerView.OnScrollListener
 import com.depromeet.presentation.R
 import com.depromeet.presentation.databinding.ActivitySeatRecordBinding
 import com.dpm.core.base.BaseActivity
 import com.dpm.core.state.UiState
+import com.dpm.designsystem.SpotDropDownSpinner
 import com.dpm.designsystem.SpotImageSnackBar
 import com.dpm.domain.entity.response.home.ResponseMySeatRecord
+import com.dpm.domain.entity.response.home.ResponseReviewDate
 import com.dpm.presentation.extension.loadAndCircleProfile
 import com.dpm.presentation.extension.setOnSingleClickListener
 import com.dpm.presentation.home.ProfileEditActivity
-import com.dpm.presentation.seatrecord.adapter.RecordListItem
-import com.dpm.presentation.seatrecord.adapter.SeatRecordAdapter
+import com.dpm.presentation.seatrecord.adapter.DateMonthAdapter
+import com.dpm.presentation.seatrecord.adapter.MonthRecordAdapter
 import com.dpm.presentation.seatrecord.dialog.ConfirmDeleteDialog
 import com.dpm.presentation.seatrecord.dialog.RecordEditDialog
+import com.dpm.presentation.seatrecord.uiMapper.MonthReviewData
 import com.dpm.presentation.seatrecord.viewmodel.EditUi
 import com.dpm.presentation.seatrecord.viewmodel.SeatRecordViewModel
 import com.dpm.presentation.seatreview.ReviewActivity
 import com.dpm.presentation.util.CalendarUtil
-import com.dpm.presentation.util.Utils
-import com.facebook.shimmer.ShimmerFrameLayout
 import dagger.hilt.android.AndroidEntryPoint
+import timber.log.Timber
 
 @AndroidEntryPoint
 class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
@@ -45,7 +46,9 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
         const val PROFILE_CHEER_TEAM = "profile_cheer_team"
     }
 
-    private lateinit var adapter: SeatRecordAdapter
+    private lateinit var dateMonthAdapter: DateMonthAdapter
+    private lateinit var monthRecordAdapter: MonthRecordAdapter
+    private lateinit var yearAdapter: SpotDropDownSpinner<String>
     private val viewModel: SeatRecordViewModel by viewModels()
     private var isLoading: Boolean = false
 
@@ -57,8 +60,7 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
                 val nickname = data?.getStringExtra(ProfileEditActivity.PROFILE_NAME) ?: ""
                 val profileImage = data?.getStringExtra(ProfileEditActivity.PROFILE_IMAGE) ?: ""
                 val teamId = data?.getIntExtra(ProfileEditActivity.PROFILE_CHEER_TEAM_ID, 0) ?: 0
-                val teamName =
-                    data?.getStringExtra(ProfileEditActivity.PROFILE_CHEER_TEAM_NAME) ?: ""
+                val teamName = data?.getStringExtra(ProfileEditActivity.PROFILE_CHEER_TEAM_NAME)
 
                 viewModel.updateProfile(nickname, profileImage, teamId, teamName)
             }
@@ -73,10 +75,10 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
     }
 
     private fun initView() {
-        initViewStatusBar()
+        initMonthAdapter()
+        initReviewList()
         viewModel.getReviewDate()
         viewModel.getLocalProfile()
-        initRecordAdapter()
     }
 
     private fun initEvent() {
@@ -84,22 +86,21 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
             ivBack.setOnClickListener {
                 finish()
             }
-//            recordSpotAppbar.setNavigationOnClickListener {
-//                finish()
-//            }
             fabRecordUp.setOnClickListener {
-                rvSeatRecord.smoothScrollToPosition(0)
+                ssvRecord.smoothScrollTo(0, 0)
             }
             fabRecordPlus.setOnClickListener {
                 navigateToReviewActivity()
             }
-
-            /** 게시물 없을 때 **/
-            ivRecordProfile.setOnSingleClickListener { navigateToProfileEditActivity() }
-            ivRecordEdit.setOnSingleClickListener { navigateToProfileEditActivity() }
-            btRecordWriteRecord.setOnSingleClickListener { navigateToReviewActivity() }
-
-            /** 로딩 실패 **/
+            ivRecordEdit.setOnClickListener {
+                navigateToProfileEditActivity()
+            }
+            ivRecordProfile.setOnClickListener {
+                navigateToProfileEditActivity()
+            }
+            btRecordWriteRecord.setOnClickListener {
+                navigateToReviewActivity()
+            }
             btRecordFailRefresh.setOnSingleClickListener {
                 if (viewModel.reviews.value is UiState.Failure || viewModel.date.value is UiState.Failure) {
                     makeSpotImageAppbar("리뷰를 불러오는데 실패하였습니다.")
@@ -112,135 +113,48 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
             csbvHelpInfo.setOnClickListener {
                 csbvHelpInfo.visibility = GONE
             }
+            tvSeatView.setOnSingleClickListener {
+                vSeatViewDivider.visibility = VISIBLE
+                vIntuitiveReviewDivider.visibility = GONE
+            }
+            tvIntuitiveReview.setOnSingleClickListener {
+                vSeatViewDivider.visibility = GONE
+                vIntuitiveReviewDivider.visibility = VISIBLE
+            }
         }
     }
 
     private fun initObserver() {
         observeDates()
         observeReviews()
-        observeEvents()
         observeProfile()
+        observeEvents()
     }
-
-    private fun initViewStatusBar() {
-        Utils(this).apply {
-            setStatusBarColor(window, com.depromeet.designsystem.R.color.color_background_tertiary)
-            setBlackSystemBarIconColor(window)
-        }
-    }
-
-    private fun initRecordAdapter() {
-        adapter = SeatRecordAdapter(
-            reviewClick = {
-                viewModel.setClickedReviewId(it.id)
-                supportFragmentManager.commit {
-                    replace(
-                        R.id.fcv_record,
-                        SeatDetailRecordFragment(),
-                        SeatDetailRecordFragment.SEAT_RECORD_TAG
-                    )
-                    addToBackStack(null)
-                }
-            },
-            monthClick = { month ->
-                viewModel.setSelectedMonth(month)
-                binding.rvSeatRecord.smoothScrollToPosition(0)
-            },
-            yearClick = { year ->
-                viewModel.setSelectedYear(year)
-                binding.rvSeatRecord.smoothScrollToPosition(0)
-            },
-            profileEditClick = {
-                navigateToProfileEditActivity()
-            },
-            reviewEditClick = { reviewId ->
-                viewModel.setEditReviewId(reviewId)
-                RecordEditDialog.newInstance(SEAT_RECORD_TAG)
-                    .apply { show(supportFragmentManager, this.tag) }
-            },
-            seatViewClick = {
-
-            },
-            intuitiveViewClick = {
-
-            }
-        )
-
-        with(binding) {
-            rvSeatRecord.adapter = adapter
-            rvSeatRecord.addItemDecoration(
-                HeaderItemDecoration(
-                    binding.rvSeatRecord,
-                    shouldFadeOutHeader = true,
-                    isHeader = { position -> (rvSeatRecord.adapter as SeatRecordAdapter).isHeader(position) }
-                )
-            )
-            rvSeatRecord.addOnScrollListener(object : OnScrollListener() {
-                override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
-                    val scrollTop = !binding.rvSeatRecord.canScrollVertically(-1)
-                    val scrollBottom = !binding.rvSeatRecord.canScrollVertically(1)
-                    if (scrollBottom && !isLoading && viewModel.reviews.value is UiState.Success) {
-                        if ((viewModel.reviews.value as UiState.Success).data.hasNext) {
-                            isLoading = true
-                            viewModel.loadNextSeatRecords()
-                        }
-                    }
-                    binding.fabRecordUp.visibility = if (scrollTop) GONE else VISIBLE
-                }
-            })
-        }
-    }
-
-    private fun View.setVisible(visible: Boolean) {
-        visibility = if (visible) VISIBLE else GONE
-    }
-
-    private fun setErrorVisibility(recordErrorType: RecordErrorType) {
-        with(binding) {
-            rvSeatRecord.setVisible(recordErrorType == RecordErrorType.NONE)
-            clRecordError.setVisible(recordErrorType != RecordErrorType.NONE)
-            clRecordEmpty.setVisible(recordErrorType == RecordErrorType.EMPTY)
-            clRecordFail.setVisible(recordErrorType == RecordErrorType.FAIL)
-            fabRecordUp.visibility = GONE
-        }
-    }
-
 
     private fun observeDates() {
         viewModel.date.asLiveData().observe(this) { state ->
             when (state) {
                 is UiState.Success -> {
-                    setErrorVisibility(RecordErrorType.NONE)
-                    stopDateShimmer()
-                    val reviewState = viewModel.reviews.value
-                    if (reviewState is UiState.Success) {
-                        adapter.updateItemAt(2, RecordListItem.Date(state.data.yearMonths))
-                    } else {
-                        adapter.submitList(
-                            listOf(
-                                RecordListItem.Profile(ResponseMySeatRecord.MyProfileResponse()),
-                                RecordListItem.ReviewType,
-                                RecordListItem.Date(state.data.yearMonths),
-                                RecordListItem.Record(emptyList())
-                            )
-                        )
-                    }
+                    setErrorVisibility(SeatRecordErrorType.NONE)
+                    setYearSpinner(state.data)
+                    dateMonthAdapter.submitList(state.data.yearMonths.first { it.isClicked }.months)
                     viewModel.getSeatRecords()
                 }
 
                 is UiState.Empty -> {
-                    setErrorVisibility(RecordErrorType.EMPTY)
-                    stopAllShimmer()
+                    setErrorVisibility(SeatRecordErrorType.EMPTY)
+                    setShimmer(false)
                 }
 
                 is UiState.Loading -> {
-                    startAllShimmer()
+                    setShimmer(true)
                 }
 
                 is UiState.Failure -> {
-                    makeSpotImageAppbar("리뷰를 불러오는데 실패하였습니다.")
-                    setErrorVisibility(RecordErrorType.FAIL)
-                    stopAllShimmer()
+                    setErrorVisibility(SeatRecordErrorType.FAIL)
+
+                    binding.rvRecordMonthDetail.visibility = GONE
+                    setShimmer(false)
                 }
 
             }
@@ -251,67 +165,30 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
         viewModel.reviews.asLiveData().observe(this) { state ->
             when (state) {
                 is UiState.Success -> {
-                    stopReviewShimmer()
-                    val newProfileItem = RecordListItem.Profile(state.data.profile)
-                    val newDateItem =
-                        RecordListItem.Date((viewModel.date.value as UiState.Success).data.yearMonths)
-                    val newRecordItem = RecordListItem.Record(state.data.reviews)
-                    val newList = adapter.currentList.toMutableList()
-                    newList[0] = newProfileItem
-                    newList[1] = RecordListItem.ReviewType
-                    newList[2] = newDateItem
-                    newList[3] = newRecordItem
-
-                    adapter.submitList(newList)
-
+                    setShimmer(false)
+                    setProfile(state.data.profile)
+                    updateReviewList(state.data.reviews)
                     isLoading = false
-                    setErrorVisibility(RecordErrorType.NONE)
+                    setErrorVisibility(SeatRecordErrorType.NONE)
                 }
 
                 is UiState.Loading -> {
                 }
 
                 is UiState.Empty -> {
+                    setShimmer(false)
+                    monthRecordAdapter.submitList(emptyList())
                     makeSpotImageAppbar("해당 날짜에 작성한 글이 없습니다.")
                 }
 
                 is UiState.Failure -> {
                     makeSpotImageAppbar("리뷰를 불러오는데 실패하였습니다.")
-                    setErrorVisibility(RecordErrorType.FAIL)
-                    stopReviewShimmer()
+                    setShimmer(false)
+                    setErrorVisibility(SeatRecordErrorType.FAIL)
                 }
             }
         }
     }
-
-    private fun observeProfile() {
-        viewModel.profile.asLiveData().observe(this) {
-            profileNone(it)
-            //TODO : 리사이클러뷰 프로필
-        }
-    }
-
-    private fun profileNone(profile: ResponseMySeatRecord.MyProfileResponse) {
-        with(binding) {
-            if (profile.teamId != null && profile.teamId != 0) {
-                csbvRecordTitle.setTextPart(
-                    "${profile.teamName}의 Lv.", profile.level,
-                    " ${profile.levelTitle}"
-                )
-
-            } else {
-                csbvRecordTitle.setTextPart(
-                    "모두를 응원하는 Lv.", profile.level, " ${profile.levelTitle}"
-                )
-            }
-            csbvHelpInfo.setText("내 기록이 야구팬들에게 도움된 횟수에요!")
-            ivRecordProfile.loadAndCircleProfile(profile.profileImage)
-            tvRecordNickname.text = profile.nickname
-            tvRecordCount.text = "0"
-            "${CalendarUtil.getCurrentYear()}년".also { tvRecordYear.text = it }
-        }
-    }
-
 
     private fun observeEvents() {
         viewModel.deleteClickedEvent.asLiveData().observe(this) { state ->
@@ -329,12 +206,171 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
         }
     }
 
-    private fun navigateToProfileEditActivity() {
-        editProfileLauncher.launch(Intent(this, ProfileEditActivity::class.java).apply {
-            putExtra(PROFILE_NAME, viewModel.profile.value.nickname)
-            putExtra(PROFILE_IMAGE, viewModel.profile.value.profileImage)
-            putExtra(PROFILE_CHEER_TEAM, viewModel.profile.value.teamId)
+    private fun observeProfile() {
+        viewModel.profile.asLiveData().observe(this) {
+            setProfile(it)
+        }
+    }
+
+    private fun setProfile(data: ResponseMySeatRecord.MyProfileResponse) {
+        with(binding) {
+            if (data.teamId != null && data.teamId != 0) {
+                csbvRecordTitle.setTextPart(
+                    "${data.teamName}의 Lv.", data.level,
+                    " ${data.levelTitle}"
+                )
+
+            } else {
+                csbvRecordTitle.setTextPart(
+                    "모두를 응원하는 Lv.", data.level, " ${data.levelTitle}"
+                )
+            }
+            csbvHelpInfo.setText("내 기록이 야구팬들에게 도움된 횟수에요!")
+            ivRecordProfile.loadAndCircleProfile(data.profileImage)
+            tvRecordNickname.text = data.nickname
+            tvRecordCount.text = "0"
+            tvRecordCount.text = data.reviewCount.toString()
+        }
+    }
+
+    private fun setYearSpinner(data: ResponseReviewDate) {
+        val years = data.yearMonths.map { it.year }
+        val yearList = years.map { "${it}년" }
+        val selectedYear = data.yearMonths.firstOrNull { it.isClicked }?.year
+        val selectedPosition = years.indexOf(selectedYear).takeIf { it >= 0 } ?: 0
+
+
+        if (!::yearAdapter.isInitialized) {
+            yearAdapter = SpotDropDownSpinner(yearList, selectedPosition)
+            binding.spinnerRecordYear.adapter = yearAdapter
+
+
+            binding.spinnerRecordYear.setSelection(selectedPosition)
+            binding.spinnerRecordYear.onItemSelectedListener =
+                object : AdapterView.OnItemSelectedListener {
+                    override fun onItemSelected(
+                        parent: AdapterView<*>?,
+                        view: View?,
+                        position: Int,
+                        id: Long,
+                    ) {
+                        yearAdapter.setSelectedItemPosition(position)
+                        viewModel.setSelectedYear(years[position])
+                        binding.ssvRecord.smoothScrollTo(0, 0)
+                    }
+
+                    override fun onNothingSelected(p0: AdapterView<*>?) {}
+                }
+
+        } else {
+            yearAdapter.updateData(yearList, selectedPosition)
+            binding.spinnerRecordYear.setSelection(selectedPosition)
+        }
+    }
+
+    private fun initMonthAdapter() {
+        if (!::dateMonthAdapter.isInitialized) {
+            dateMonthAdapter = DateMonthAdapter(
+                monthClick = { month ->
+                    viewModel.setSelectedMonth(month)
+                    binding.ssvRecord.smoothScrollTo(0, 0)
+                }
+            )
+            binding.rvRecordMonth.adapter = dateMonthAdapter
+        }
+    }
+
+    private fun View.setVisible(visible: Boolean) {
+        visibility = if (visible) VISIBLE else GONE
+    }
+
+    private fun setErrorVisibility(errorType: SeatRecordErrorType) {
+        with(binding) {
+            rvRecordMonthDetail.setVisible(errorType == SeatRecordErrorType.NONE)
+            tvErrorMonth.setVisible(errorType != SeatRecordErrorType.NONE)
+            tvErrorYear.setVisible(errorType != SeatRecordErrorType.NONE)
+            "${CalendarUtil.getCurrentYear()}년".also { tvErrorYear.text = it }
+            clRecordEmpty.setVisible(errorType == SeatRecordErrorType.EMPTY)
+            clRecordFail.setVisible(errorType == SeatRecordErrorType.FAIL)
+            fabRecordUp.visibility = GONE
+            if (errorType == SeatRecordErrorType.NONE) ssvRecord.header =
+                binding.clRecordStickyHeader
+        }
+    }
+
+    private fun initReviewList() {
+        monthRecordAdapter = MonthRecordAdapter()
+        binding.rvRecordMonthDetail.adapter = monthRecordAdapter
+        binding.rvRecordMonthDetail.itemAnimator = null
+
+        monthRecordAdapter.itemRecordClickListener =
+            object : MonthRecordAdapter.OnItemRecordClickListener {
+
+                override fun onItemRecordClick(item: ResponseMySeatRecord.ReviewResponse) {
+                    viewModel.setClickedReviewId(item.id)
+                    supportFragmentManager.commit {
+                        replace(
+                            R.id.fcv_record,
+                            SeatDetailRecordFragment(),
+                            SeatDetailRecordFragment.SEAT_RECORD_TAG
+                        )
+                        addToBackStack(null)
+                    }
+                }
+
+                override fun onMoreRecordClick(reviewId: Int) {
+                    viewModel.setEditReviewId(reviewId)
+                    RecordEditDialog.newInstance(SEAT_RECORD_TAG)
+                        .apply { show(supportFragmentManager, this.tag) }
+                }
+
+            }
+
+
+        //TODO 이렇게 무한스크롤 하면 되네..
+        binding.ssvRecord.setOnScrollChangeListener(NestedScrollView.OnScrollChangeListener { v, _, scrollY, _, _ ->
+            if (scrollY == (v.getChildAt(0).measuredHeight - v.measuredHeight)) {
+                Timber.d("test 끝 도달")
+                val hasNextPage =
+                    (viewModel.reviews.value as? UiState.Success)?.data?.hasNext == true
+                if (hasNextPage && !isLoading) {
+                    isLoading = true
+                    viewModel.loadNextSeatRecords()
+                }
+            }
+            binding.fabRecordUp.visibility = if (scrollY == 0) GONE else VISIBLE
         })
+    }
+
+
+    private fun updateReviewList(reviews: List<ResponseMySeatRecord.ReviewResponse>) {
+        val groupList =
+            reviews.groupBy { CalendarUtil.getMonthFromDateFormat(it.date) }
+                .map { (month, reviews) ->
+                    MonthReviewData(month, reviews)
+                }
+        monthRecordAdapter.submitList(groupList)
+    }
+
+    private fun navigateToProfileEditActivity() {
+        val currentState = viewModel.reviews.value
+        if (currentState is UiState.Success) {
+            editProfileLauncher.launch(Intent(this, ProfileEditActivity::class.java).apply {
+                with(currentState.data) {
+                    putExtra(PROFILE_NAME, this.profile.nickname)
+                    putExtra(PROFILE_IMAGE, this.profile.profileImage)
+                    putExtra(PROFILE_CHEER_TEAM, this.profile.teamId)
+                }
+            })
+        }
+    }
+
+    private fun navigateToReviewActivity() {
+        Intent(this@SeatRecordActivity, ReviewActivity::class.java).apply {
+            startActivity(
+                this
+            )
+        }
     }
 
     private fun moveConfirmationDialog() {
@@ -344,23 +380,16 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
 
 
     private fun moveEditReview() {
-        makeSpotImageAppbar("수정은 추후에 열릴 예정입니다!")
-//        viewModel.setEditReview(viewModel.editReviewId.value)
-//        supportFragmentManager.commit {
-//            replace(
-//                R.id.fcv_record,
-//                EditReviewFragment(),
-//                EditReviewFragment.EDIT_REIVIEW_TAG
-//            )
-//            addToBackStack(null)
-//        }
+
     }
 
-    private fun navigateToReviewActivity() {
-        Intent(this@SeatRecordActivity, ReviewActivity::class.java).apply {
-            startActivity(
-                this
-            )
+    private fun setShimmer(isLoading: Boolean) = with(binding) {
+        if (isLoading) {
+            shimmerRecord.startShimmer()
+            shimmerRecord.visibility = VISIBLE
+        } else {
+            shimmerRecord.stopShimmer()
+            shimmerRecord.visibility = GONE
         }
     }
 
@@ -375,46 +404,7 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
         ).show()
     }
 
-    private fun startShimmerWithVisibility(shimmerView: ShimmerFrameLayout) {
-        shimmerView.apply {
-            visibility = VISIBLE
-            startShimmer()
-        }
-    }
-
-    private fun stopShimmerWithVisibility(shimmerView: ShimmerFrameLayout) {
-        shimmerView.apply {
-            stopShimmer()
-            visibility = INVISIBLE
-        }
-    }
-
-    private fun startAllShimmer() {
-        with(binding) {
-            startShimmerWithVisibility(shimmerProfile)
-            startShimmerWithVisibility(shimmerDate)
-            startShimmerWithVisibility(shimmerReview)
-        }
-    }
-
-    private fun stopAllShimmer() {
-        with(binding) {
-            stopShimmerWithVisibility(shimmerProfile)
-            stopShimmerWithVisibility(shimmerDate)
-            stopShimmerWithVisibility(shimmerReview)
-        }
-    }
-
-    private fun stopDateShimmer() {
-        stopShimmerWithVisibility(binding.shimmerDate)
-    }
-
-    private fun stopReviewShimmer() {
-        stopShimmerWithVisibility(binding.shimmerProfile)
-        stopShimmerWithVisibility(binding.shimmerReview)
-    }
-
-    enum class RecordErrorType {
+    enum class SeatRecordErrorType {
         EMPTY,
         FAIL,
         NONE
